requirements.txt
numpy
pandas
matplotlib
seaborn
ipykernel
jinja2
python-dotenv